# api-doc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job API",
        "version": "0.1"
    },
    "paths": {
        "/cronjob": {
            "get": {
                "tags": [
                    "CronJob"
                ],
                "description": "Returns a list of cronjobs",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CronJob"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "In case invalid parameters were provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "In case of an internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CronJob"
                ],
                "description": "Creates a cronjob",
                "operationId": "Abc\\Job\\Controller\\CronJobController::create",
                "requestBody": {
                    "description": "CronJob object to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CronJob"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CronJob"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "In case invalid parameters were provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "In case of an internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cronjob/{id}": {
            "get": {
                "tags": [
                    "CronJob"
                ],
                "description": "Returns a cronjob",
                "operationId": "Abc\\Job\\Controller\\CronJobController::find",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The unique id of the cronjob",
                        "required": true,
                        "schema": {
                            "$ref": "#components/schemas/CronJob/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CronJob"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "In case invalid parameters were provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "In case a cronjob with the given id is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "In case of an internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "CronJob"
                ],
                "description": "Updates a cronjob",
                "operationId": "Abc\\Job\\Controller\\CronJobController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The unique id of the cronjob",
                        "required": true,
                        "schema": {
                            "$ref": "#components/schemas/CronJob/properties/id"
                        }
                    }
                ],
                "requestBody": {
                    "description": "CronJob object to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CronJob"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CronJob"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "In case invalid parameters were provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "In case a cronjob with the given id is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "In case of an internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CronJob"
                ],
                "description": "Deletes a cronjob",
                "operationId": "Abc\\Job\\Controller\\CronJobController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The unique id of the cronjob",
                        "required": true,
                        "schema": {
                            "$ref": "#components/schemas/CronJob/properties/id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "In case a cronjob with the given id is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "In case of an internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProblem"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CronJob": {
                "description": "A job that is processed according to cronjob expression",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": "true"
                    },
                    "schedule": {
                        "description": "The schedule in Cron format, see https://en.wikipedia.org/wiki/Cron.",
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/Job/properties/type"
                    },
                    "name": {
                        "$ref": "#/components/schemas/Job/properties/name"
                    },
                    "input": {
                        "$ref": "#/components/schemas/Job/properties/input"
                    },
                    "externalId": {
                        "$ref": "#/components/schemas/Job/properties/externalId"
                    },
                    "children": {
                        "$ref": "#/components/schemas/Job/properties/children"
                    },
                    "updated": {
                        "description": "The datetime the cronjob was updated",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "created": {
                        "description": "The datetime the cronjob was created",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
            "ApiProblem": {
                "description": "Problem Details for HTTP APIs according to RFC7 807",
                "properties": {
                    "type": {
                        "description": " A URI reference [RFC3986] that identifies the problem type.",
                        "type": "string"
                    },
                    "title": {
                        "description": "A short, human-readable summary of the problem type.",
                        "type": "string"
                    },
                    "status": {
                        "description": "The HTTP status code ([RFC7231], Section 6) generated by the origin server for this occurrence of the problem.",
                        "type": "integer"
                    },
                    "detail": {
                        "description": "A human-readable explanation specific to this occurrence of the problem.",
                        "type": "string"
                    },
                    "instance": {
                        "description": "A URI reference that identifies the specific occurrence of the problem.",
                        "type": "string"
                    },
                    "invalidParams": {
                        "type": "array",
                        "items": {
                            "title": "InvalidParameter",
                            "properties": {
                                "name": {
                                    "description": "The name or path of the invalid property",
                                    "type": "string"
                                },
                                "reason": {
                                    "description": "A reason explaining the validation constraint violation",
                                    "type": "string"
                                },
                                "valie": {
                                    "description": "The value that was provided",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
...
    }
}